requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

# ВАЖНО: корректный httpx под PTB 20.x
httpx==0.26.0

aiohttp==3.9.5
aiofiles==23.2.1
redis==5.0.6
SQLAlchemy==2.0.32
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.7.4
python-dotenv==1.0.1

# Telegram (если используешь PTB; aiogram можно оставить, но он не обязателен)
python-telegram-bot==20.8

# tests & lint
pytest==8.3.2
pytest-asyncio==0.23.8
respx==0.21.1
flake8==7.1.1
pydantic-settings>=2.2,<3

